+++ src/test/java/com/codeborne/selenide/BrowserTest.java
@@ -7,7 +7,6 @@ import static com.codeborne.selenide.Browsers.EDGE;
 import static com.codeborne.selenide.Browsers.FIREFOX;
 import static com.codeborne.selenide.Browsers.IE;
 import static com.codeborne.selenide.Browsers.INTERNET_EXPLORER;
-import static com.codeborne.selenide.Browsers.LEGACY_FIREFOX;
 import static com.codeborne.selenide.Browsers.OPERA;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -24,7 +23,6 @@ final class BrowserTest {
   void mostBrowsersSupportInsecureCerts() {
     assertThat(new Browser(CHROME, false).supportsInsecureCerts()).isTrue();
     assertThat(new Browser(FIREFOX, false).supportsInsecureCerts()).isTrue();
-    assertThat(new Browser(LEGACY_FIREFOX, false).supportsInsecureCerts()).isTrue();
     assertThat(new Browser(OPERA, false).supportsInsecureCerts()).isTrue();
   }