+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## 6.0.0 (planned to 1x.10.2021)
+## 6.0.0 (planned to 25.10.2021)
 * upgrade to Selenium 4.0.0
 * #1581 report exact actual value at the moment of failure
 * Remove lots of deprecated methods  --  thanks @BorisOsipov for PR https://github.com/selenide/selenide/pull/1607/files
+* Remove support for "legacy_firefox" (upto ESR 52)
 * Make Selenide.sleep(N) guarantee the sleep duration
 
 ## 5.25.0 (released 28.09.2021)